+++ FreeGroup/main/profile_reduce.cpp
@@ -19,7 +19,7 @@
 
 using crag::slp::Vertex;
 using crag::slp::PreorderInspector;
-typedef crag::slp::TerminalVertexTemplate<int> TerminalVertex;
+using crag::slp::TerminalVertex;
 using crag::slp::NonterminalVertex;
 using crag::slp::MatchingTable;
 using crag::UniformAutomorphismSLPGenerator;
@@ -31,11 +31,11 @@ int main() {
   const size_t RANK = 3;
   const size_t ENDOMORPHISMS_NUMBER = 100;
   size_t seed = 112233;
-  UniformAutomorphismSLPGenerator<int> generator(RANK, seed);
+  UniformAutomorphismSLPGenerator<> generator(RANK, seed);
   auto begin = std::chrono::system_clock::now();
   int count = REPEAT;
   while (--count >= 0) {
-    auto image = EndomorphismSLP<int>::composition(ENDOMORPHISMS_NUMBER, generator).image(1);
+    auto image = EndomorphismSLP::composition(ENDOMORPHISMS_NUMBER, generator).image(1);
 
     Vertex reduced = reduce(image);
     auto end = std::chrono::system_clock::now();
