+++ FreeGroup/test/slp_reduce.cpp
@@ -24,11 +24,12 @@ namespace crag {
 namespace slp {
 namespace {
 
-typedef TerminalVertexTemplate<char> TerminalVertex;
+constexpr TerminalSymbol terminal_a = TerminalSymbol{} + 1;
+constexpr TerminalSymbol terminal_b = TerminalSymbol{} + 2;
 
 Vertex get_random_slp_on_2_letters(unsigned int WORD_SIZE) {
-  TerminalVertex a('a');
-  TerminalVertex b('b');
+  TerminalVertex a(terminal_a);
+  TerminalVertex b(terminal_b);
 
   int random_word = rand() % (1 << WORD_SIZE);
   std::vector<unsigned int> random_word_split;
@@ -58,11 +59,11 @@ Vertex get_random_slp_on_2_letters(unsigned int WORD_SIZE) {
 //  int repeat = 1;
 //  while (--repeat >= 0) {
 //    Vertex slp = get_random_slp_on_2_letters(size);
-//    std::string slp_string(VertexWord<char>(slp).begin(), VertexWord<char>(slp).end());
+//    std::string slp_string(VertexWord(slp).begin(), VertexWord(slp).end());
 //    for (size_t start = 1; start < size; ++start) {
 //      for (size_t end = start + 1; end <= size; ++end) {
 //        Vertex sub_slp = get_sub_slp(slp, start, end);
-//        std::string subslp_string(VertexWord<char>(sub_slp).begin(), VertexWord<char>(sub_slp).end());
+//        std::string subslp_string(VertexWord(sub_slp).begin(), VertexWord(sub_slp).end());
 //        ASSERT_EQ(slp_string.substr(start, end - start), subslp_string) << slp_string << '[' << start << ':' <<end << ']';
 //      }
 //    }
@@ -75,7 +76,7 @@ TEST(SubSLP, StressTest) {
   srand(time(0));
   while (--repeat >= 0) {
     Vertex slp = get_random_slp_on_2_letters(size);
-    std::string slp_string(VertexWord<char>(slp).begin(), VertexWord<char>(slp).end());
+    std::string slp_string(VertexWord(slp).begin(), VertexWord(slp).end());
     size_t start = 0;
     size_t end = 0;
     while (start == end) {
@@ -86,7 +87,7 @@ TEST(SubSLP, StressTest) {
       }
     }
     Vertex sub_slp = get_sub_slp(slp, start, end);
-    std::string subslp_string(VertexWord<char>(sub_slp).begin(), VertexWord<char>(sub_slp).end());
+    std::string subslp_string(VertexWord(sub_slp).begin(), VertexWord(sub_slp).end());
     ASSERT_EQ(slp_string.substr(start, end - start), subslp_string) << slp_string << '[' << start << ':' <<end << ']';
 
     PreorderInspector inspector(sub_slp);
@@ -99,16 +100,16 @@ TEST(SubSLP, StressTest) {
 }
 
 TEST(Reduce, Simple1) {
-  TerminalVertex a('a');
-  TerminalVertex a_(-'a');
+  TerminalVertex a(terminal_a);
+  TerminalVertex a_(-terminal_a);
   NonterminalVertex mNull(a, a_);
 
   EXPECT_EQ(Vertex(), reduce(mNull));
 }
 
 TEST(Reduce, Simple2) {
-  TerminalVertex a('a');
-  TerminalVertex b('b');
+  TerminalVertex a(terminal_a);
+  TerminalVertex b(terminal_b);
   NonterminalVertex ab(a, b);
   NonterminalVertex b_1a_1(b.negate(), a.negate());
   NonterminalVertex mNull(ab, b_1a_1);
@@ -159,21 +160,21 @@ std::string print_tree_preorder_single(const Vertex& vertex) {
 
 
 TEST(Reduce, Example1) {
-  TerminalVertexTemplate<int> v1(1);
-  TerminalVertexTemplate<int> v2(2);
+  TerminalVertex v1(1);
+  TerminalVertex v2(2);
   NonterminalVertex v9(v1, v2);
   Vertex v12 = NonterminalVertex(v2, v9).negate();
   NonterminalVertex v13(v2, v12);
   auto reduced = reduce(v13);
   int previous = 0;
-  for (auto symbol : VertexWord<int>(reduced)) {
+  for (auto symbol : VertexWord(reduced)) {
     EXPECT_NE(-previous, symbol) << print_tree_preorder(reduced);
     previous = symbol;
   }
 }
-//TerminalVertexTemplate<int> t1(1);
-//TerminalVertexTemplate<int> t2(2);
-//TerminalVertexTemplate<int> t3(3);
+//TerminalVertex t1(1);
+//TerminalVertex t2(2);
+//TerminalVertex t3(3);
 //NonterminalVertex v1(t1, t2);
 //NonterminalVertex v2(v1, t3);
 //NonterminalVertex v3(t1, v2);
@@ -188,9 +189,9 @@ TEST(Reduce, Example1) {
 //std::vector<int> correct = {-3, -2, -1, -3, -2, -1, 2, 3, 3, 3};
 
 TEST(Reduce, Example2) {
-  TerminalVertexTemplate<int> t1(1);
-  TerminalVertexTemplate<int> t2(2);
-  TerminalVertexTemplate<int> t3(3);
+  TerminalVertex t1(1);
+  TerminalVertex t2(2);
+  TerminalVertex t3(3);
   NonterminalVertex v1(t1, t2);
   NonterminalVertex v2(v1, t3);
   NonterminalVertex v3(t1, v2);
@@ -205,7 +206,7 @@ TEST(Reduce, Example2) {
   auto reduced = reduce(v8);
   ASSERT_EQ(correct.size(), reduced.length());
   auto correct_symbol = correct.begin();
-  for (auto symbol : VertexWord<int>(reduced)) {
+  for (auto symbol : VertexWord(reduced)) {
     ASSERT_EQ(*correct_symbol, symbol);
     ++correct_symbol;
   }
@@ -218,13 +219,13 @@ TEST(Reduce, StressTest) {
   const size_t ENDOMORPHISMS_NUMBER = 20;
   size_t seed = 0;
   while (++seed <= REPEAT) {
-    UniformAutomorphismSLPGenerator<int> generator(RANK, seed);
-    auto image = EndomorphismSLP<int>::composition(ENDOMORPHISMS_NUMBER, generator).image(1);
+    UniformAutomorphismSLPGenerator<> generator(RANK, seed);
+    auto image = EndomorphismSLP::composition(ENDOMORPHISMS_NUMBER, generator).image(1);
 
     Vertex reduced = reduce(image);
 
     std::vector<int> reduced_image;
-    for (auto symbol : VertexWord<int>(reduced)) {
+    for (auto symbol : VertexWord(reduced)) {
       if (!reduced_image.empty() && symbol == -reduced_image.back()) {
         reduced_image.pop_back();
       } else {
@@ -234,12 +235,12 @@ TEST(Reduce, StressTest) {
     std::ostringstream reduced_image_string;
     std::copy(reduced_image.begin(), reduced_image.end(), std::ostream_iterator<int>(reduced_image_string, ""));
     auto correct_symbol = reduced_image.begin();
-    for (auto symbol : VertexWord<int>(reduced)) {
+    for (auto symbol : VertexWord(reduced)) {
       ASSERT_EQ(*correct_symbol, symbol) << seed << std::endl
           << print_tree_preorder_single(image) << std::endl
           << print_tree_preorder(reduced) << std::endl
-          << VertexWord<int>(image) << std::endl
-          << VertexWord<int>(reduced) << std::endl
+          << VertexWord(image) << std::endl
+          << VertexWord(reduced) << std::endl
           << reduced_image_string.str() << std::endl;
       ++correct_symbol;
     }
@@ -251,9 +252,9 @@ TEST(Reduce, StressTest) {
 //  const size_t RANK = 3;
 //  const size_t ENDOMORPHISMS_NUMBER = 100;
 //  size_t seed = 112233;///time(0);//
-//  UniformAutomorphismSLPGenerator<int> generator(RANK, seed);
+//  UniformAutomorphismSLPGenerator<> generator(RANK, seed);
 //  while (--REPEAT >= 0) {
-//    auto image = EndomorphismSLP<int>::composition(ENDOMORPHISMS_NUMBER, generator).slp(1);
+//    auto image = EndomorphismSLP::composition(ENDOMORPHISMS_NUMBER, generator).slp(1);
 //
 //    Vertex reduced = reduce(image);
 //    std::cout << image.length() << std::endl;
